// types/films.interface.ts
export interface IFilm {
  id: number;
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: number[];
  planets: number[];
  starships: number[];
  vehicles: number[];
  species: number[];
  created: string;
  edited: string;
  url: string;
}

export interface IFilmsResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: IFilm[];
}